update nextauth config
